refactor: [M3-9616] - Move `ramda` dependent utils (#11913)
* Move isNilOrEmpty

* Move createDevicesFromStrings

* Resolve TS errors in createDevicesFromStrings.text.ts

* Move createStringsFromDevices

* Move maybeCastToNumber

* Consolidate imports

* Fix TS error

* Added changeset: Move ramda dependent utils

* Keep DevicesAsStrings type import separate for better organization

* Added changeset: Migrated ramda dependent utils to @linode/utilities package

* Avoid importing from @linode/utilities from within the same package

* Update packages/manager/.changeset/pr-11913-added-1742915642212.md


* Update packages/manager/.changeset/pr-11913-added-1742915642212.md


* Move utils changeset to correct spot

* upcoming: [M3-9534] - Initial VPC Support in the `Add Network Interface` Drawer (#11887)

* initial vpc and subnet select

* add a shared firewallselect component

* use firewall select globally

* add jsdoc comments

* add some testing

* finish up testing for now

* Added changeset: Add VPC support to the Add Network Interface Drawer

* Added changeset: Added `FirewallSelect` component

* Added changeset: Add test for Add Linode Interface drawer

* clean up changesets

* support default chips in the Firewall Select

* fix spacing regression

* properly handle disableClearable in the new Firewall Select

* support default firewalls in the Add Interface drawer

* use newer copy @coliu-akamai

* fix unit test after UX tooltip changes

---------


* test: [M3-9486, M3-9487, M3-9557] - Allow Linode create tests to pass in alternative environments (#11886)

* Delete redundant Linode Create SSH key test

* Add "env:premiumPlans" test tag

* Apply "env:premiumPlans" tag to Linode premium plan e2e test

* Only require "Premium Plans" region capability for Premium Plans Linode create test

* refactor: [M3-9617] - Move `doesRegionSupportFeature` to `utilities` package (#11891)

* Move `doesRegionSupporFeature` to `utilities` pkg

* Added changeset: Move `doesRegionSupportFeature` from `manager` to `utilities` package

* Added changeset: Move `doesRegionSupportFeature` from `manager` to `utilities` package

* refactor: [M3-8247] - Remove ramda from Utilities (#11861)

* refactor: [M3-8247] - Remove ramda from Utilities

* Add changeset

* Add changeset

* updated comment

* increase coverage for isNilorEmpty()

* feat: [UIE-8600] - IAM RBAC: add new drawer for unassigning role flow (#11893)

* feat: [UIE-8600] - IAM RBAC: add new drawer for unassigning role flow

* Added changeset: Add a new confirmation dialog for the unassigning role flow in IAM

* fix the chip's color for dark theme

* fix conflict and small improvements

* upcoming: [UIE-8515] - DBaaS: Advanced Configuration - Drawer with existing configs (#11812)

* feat: [UIE-8515] - DBaaS: Advanced Configuration - Drawer with existing configs

* feat: [UIE-8515] - update mock data

* Added changeset: DBaaS Advanced Configurations: added `getDatabaseEngineConfig` request to fetch all advanced configurations and updated types for advanced configs

* Added changeset: DBaaS Advanced Configurations: added UI for existing engine options in the drawer

* upcoming: [UIE-8515] - review fix

* upcoming: [UIE-8515] - style update, add link

* update styles after token re-organization

---------


* Resolve merge conflict

* Move createDevicesFromStrings

* Move createStringsFromDevices

* Avoid importing from @linode/utilities from within the same package

* refactor: [M3-8247] - Remove ramda from Utilities (#11861)

* refactor: [M3-8247] - Remove ramda from Utilities

* Add changeset

* Add changeset

* updated comment

* increase coverage for isNilorEmpty()

* Move isNilOrEmpty

* Move createDevicesFromStrings

* Move createStringsFromDevices

* refactor: [M3-8247] - Remove ramda from Utilities (#11861)

* refactor: [M3-8247] - Remove ramda from Utilities

* Add changeset

* Add changeset

* updated comment

* increase coverage for isNilorEmpty()

* Fix post merge conflict issues

---------
