+++ packages/manager/.changeset/pr-11913-removed-1742847225104.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Removed
+---
+
+Move ramda dependent utils ([#11913](https://github.com/linode/manager/pull/11913))
