+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeRescue/StandardRescueDialog.tsx
@@ -15,13 +15,12 @@ import {
   Paper,
   clamp,
 } from '@linode/ui';
-import { usePrevious } from '@linode/utilities';
+import { usePrevious, createDevicesFromStrings } from '@linode/utilities';
 import { styled, useTheme } from '@mui/material/styles';
 import { useSnackbar } from 'notistack';
 import * as React from 'react';
 
 import { useEventsPollingActions } from 'src/queries/events/events';
-import { createDevicesFromStrings } from 'src/utilities/createDevicesFromStrings';
 
 import { LinodePermissionsError } from '../LinodePermissionsError';
 import { DeviceSelection } from './DeviceSelection';
@@ -29,7 +28,7 @@ import { RescueDescription } from './RescueDescription';
 
 import type { ExtendedDisk } from './DeviceSelection';
 import type { APIError } from '@linode/api-v4/lib/types';
-import type { DevicesAsStrings } from 'src/utilities/createDevicesFromStrings';
+import type { DevicesAsStrings } from '@linode/utilities';
 
 interface Props {
   linodeId: number | undefined;
