+++ packages/manager/src/utilities/formikErrorUtils.ts
@@ -1,5 +1,5 @@
 import { getAPIErrorOrDefault } from './errorUtils';
-import { isNilOrEmpty } from './isNilOrEmpty';
+import { isNilOrEmpty } from '@linode/utilities';
 
 import type { APIError } from '@linode/api-v4/lib/types';
 import type { FormikErrors } from 'formik';
