+++ packages/utilities/.changeset/pr-11913-added-1742915642212.md
@@ -1,0 +1,5 @@
+---
+"@linode/utilities": Added
+---
+
+Migrate ramda dependent utils to @linode/utilities package ([#11913](https://github.com/linode/manager/pull/11913))
