+++ packages/utilities/src/helpers/createDevicesFromStrings.ts
