+++ packages/utilities/src/helpers/createStringsFromDevices.ts
@@ -3,7 +3,7 @@ import type {
   DiskDevice,
   VolumeDevice,
 } from '@linode/api-v4/lib/linodes';
-import type { DevicesAsStrings } from 'src/utilities/createDevicesFromStrings';
+import type { DevicesAsStrings } from './createDevicesFromStrings';
 
 const rdx = (
   result: DevicesAsStrings,
