+++ packages/utilities/src/helpers/index.ts
@@ -5,6 +5,8 @@ export * from './arePropsEqual';
 export * from './arrayToList';
 export * from './breakpoints';
 export * from './capitalize';
+export * from './createDevicesFromStrings';
+export * from './createStringsFromDevices';
 export * from './deepStringTransform';
 export * from './doesRegionSupportFeature';
 export * from './downloadFile';
@@ -20,10 +22,12 @@ export * from './getNewRegionLabel';
 export * from './groupByTags';
 export * from './getDisplayName';
 export * from './initWindows';
+export * from './isNilOrEmpty';
 export * from './isNumber';
 export * from './isToday';
 export * from './link';
 export * from './manuallySetVPCConfigInterfacesToActive';
+export * from './maybeCastToNumber';
 export * from './metadata';
 export * from './minute-conversion';
 export * from './mockLocalStorage';
