+++ packages/utilities/src/helpers/isNilOrEmpty.ts
