+++ packages/utilities/src/helpers/maybeCastToNumber.ts
